+++ flowcraft/generator/engine.py
@@ -1506,6 +1506,19 @@ class NextflowGenerator:
         # Flush params json to stdout
         sys.stdout.write(json.dumps(params_json))
 
+    def export_directives(self):
+        """Export pipeline directives as a JSON to stdout
+        """
+
+        directives_json = {}
+
+        # Skip first init process
+        for p in self.processes[1:]:
+            directives_json[p.template] = p.directives
+
+        # Flush params json to stdout
+        sys.stdout.write(json.dumps(directives_json))
+
     def build(self):
         """Main pipeline builder